ref: create a uri folder
