+++ src/models/mod.rs
@@ -92,7 +92,11 @@ impl PubkyAppObject {
 
 #[cfg(test)]
 mod tests {
-    use crate::utils::*;
+    use crate::{
+        blob_uri_builder, bookmark_uri_builder, feed_uri_builder, file_uri_builder,
+        follow_uri_builder, last_read_uri_builder, mute_uri_builder, post_uri_builder,
+        tag_uri_builder, user_uri_builder,
+    };
 
     use super::*;
