+++ src/uri/mod.rs
@@ -1,0 +1,9 @@
+//! Pubky URI parsing and construction for `pubky.app` resources.
+
+mod builders;
+mod parsed;
+mod resource;
+
+pub use builders::*;
+pub use parsed::ParsedUri;
+pub use resource::Resource;
