+++ src/uri/parsed.rs
@@ -6,64 +6,9 @@ use crate::{
 };
 use serde::{Deserialize, Serialize};
 use std::convert::TryFrom;
-use std::fmt;
 use url::Url;
 
-#[derive(Debug, PartialEq, Default, Clone, Serialize, Deserialize)]
-pub enum Resource {
-    User,
-    Post(String),
-    Follow(PubkyId),
-    Mute(PubkyId),
-    Bookmark(String),
-    Tag(String),
-    File(String),
-    Blob(String),
-    Feed(String),
-    LastRead,
-    #[default]
-    Unknown,
-}
-
-impl fmt::Display for Resource {
-    /// Returns the resource name without any identifier.
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // Use the associated constant for each resource type, trimming any trailing '/'
-        let name = match self {
-            Resource::User => PubkyAppUser::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::LastRead => PubkyAppLastRead::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Post(_) => PubkyAppPost::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Follow(_) => PubkyAppFollow::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Mute(_) => PubkyAppMute::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Bookmark(_) => PubkyAppBookmark::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Tag(_) => PubkyAppTag::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::File(_) => PubkyAppFile::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Blob(_) => PubkyAppBlob::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Feed(_) => PubkyAppFeed::PATH_SEGMENT.trim_end_matches('/'),
-            Resource::Unknown => "unknown",
-        };
-        write!(f, "{}", name)
-    }
-}
-
-impl Resource {
-    /// Returns the identifier as a `Some(String)` if the resource variant holds one,
-    /// or `None` if there is no identifier.
-    pub fn id(&self) -> Option<String> {
-        match self {
-            Resource::Post(id) => Some(id.clone()),
-            Resource::Follow(id) => Some(id.to_string()),
-            Resource::Mute(id) => Some(id.to_string()),
-            Resource::Bookmark(id) => Some(id.clone()),
-            Resource::Tag(id) => Some(id.clone()),
-            Resource::File(id) => Some(id.clone()),
-            Resource::Blob(id) => Some(id.clone()),
-            Resource::Feed(id) => Some(id.clone()),
-            // The following variants do not carry an id.
-            Resource::User | Resource::LastRead | Resource::Unknown => None,
-        }
-    }
-}
+use super::Resource;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ParsedUri {
@@ -75,8 +20,6 @@ impl ParsedUri {
     /// Converts the [ParsedUri] back into its URI string representation.
     /// Returns an error if the resource is Unknown.
     pub fn try_to_uri_str(&self) -> Result<String, String> {
-        use crate::traits::{HasIdPath, HasPath};
-
         let path = match &self.resource {
             Resource::User => PubkyAppUser::create_path(),
             Resource::LastRead => PubkyAppLastRead::create_path(),
@@ -181,7 +124,11 @@ impl TryFrom<String> for ParsedUri {
 
 #[cfg(test)]
 mod tests {
-    use crate::utils::*;
+    use crate::{
+        blob_uri_builder, bookmark_uri_builder, feed_uri_builder, file_uri_builder,
+        follow_uri_builder, last_read_uri_builder, mute_uri_builder, post_uri_builder,
+        tag_uri_builder, user_uri_builder,
+    };
 
     use super::*;
